bot/pyproject.toml
[tool.poetry]
name = "morphobot-bot"
version = "0.1.0"
description = "Frontend Telegram service for morphobot"
authors = ["Stankevich Andrey <stankevich.as@phystech.edu>"]
license = "Apache 2.0"
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.0"
python-telegram-bot = { extras = ["webhooks"], version = "^22.3" }
httpx = "^0.27.0"
pydantic = "^2.8"
flake8 = "^7.3.0"
flake8-pyproject = "^1.2.3"
flake8-bugbear = "^24.12.12"
flake8-import-order = "^0.19.2"
flake8-docstrings = "^1.7.0"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.21"
black = "^24.0.0"
mypy = "^1.0.0"
pillow = "^11.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
